feat: reorder sale-product price block and cast uplift to float
For products on sale, the HTML block now renders in this order:
1. Regular price (gray, struck through)
2. Sale/transfer price (red, prominent)
3. "Precio por pago con transferencia" caption
4. Other-payment price = sale price × (1 + uplift) (gray, smaller)
5. Installments line

For products NOT on sale the layout is unchanged.

Also casts WPP_Settings::get('uplift') to (float) at all three
call sites in class-checkout-fee.php as a defensive coding practice.

+++ includes/class-checkout-fee.php
@@ -62,7 +62,7 @@ class WPP_Checkout_Fee {
 			return;
 		}
 
-		$uplift   = WPP_Settings::get( 'uplift' );
+		$uplift   = (float) WPP_Settings::get( 'uplift' );
 		$subtotal = $cart->get_subtotal();
 
 		if ( $subtotal <= 0 || $uplift <= 0 ) {
@@ -187,7 +187,7 @@ class WPP_Checkout_Fee {
 			return $cart_subtotal_html;
 		}
 
-		$uplift = WPP_Settings::get( 'uplift' );
+		$uplift = (float) WPP_Settings::get( 'uplift' );
 		if ( $uplift <= 0 ) {
 			return $cart_subtotal_html;
 		}
@@ -251,7 +251,7 @@ class WPP_Checkout_Fee {
 			return $fallback_html;
 		}
 
-		$uplift = WPP_Settings::get( 'uplift' );
+		$uplift = (float) WPP_Settings::get( 'uplift' );
 		if ( $uplift <= 0 ) {
 			return $fallback_html;
 		}
